added starting data

+++ src/main/java/ru/itpark/larec/dto/ProductDto.java
@@ -4,16 +4,14 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class ProductDto {
     private int experimentId;
     private String name;
-    private BigDecimal molecularWeight;
+    private double molecularWeight;
     private String formula;
     private String linkForDescription;
-    private BigDecimal amountGram;
+    private double amountGram;
 }
