+++ src/main/java/ru/itpark/larec/dto/ReagentDto.java
@@ -1,11 +1,8 @@
 package ru.itpark.larec.dto;
 
-
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
@@ -14,8 +11,8 @@ public class ReagentDto {
 
     private int experimentId;
     private String name;
-    private BigDecimal molecularWeight;
+    private double molecularWeight;
     private String formula;
     private String linkForDescription;
-    private BigDecimal amountGram;
+    private double amountGram;
 }
